// @flywave/flywave-topo/src/common/render/primitives/mesh-params.d.ts
import { type AuxChannelTable } from "./aux-channel-table";
import { type EdgeParams } from "./edge-params";
import { type SurfaceParams } from "./surface-params";
import { type VertexTable } from "./vertex-table";
export interface MeshParams {
    vertices: VertexTable;
    surface: SurfaceParams;
    edges?: EdgeParams;
    isPlanar: boolean;
    auxChannels?: AuxChannelTable;
}